An unknown owner ID is properly rejected by the REST API
Currently when you try to update the owner ID via the REST routes
PUT /docman_embedded_files/{id}/metadata or PUT /docman_files/{id}/metadata
with an incorrect ID (e.g. with 0) you get an incorrect error message
and you might get an incorrect HTTP status (403 instead of 400).

This contribution makes sure the returned HTTP status code is 400 and
that the error message matches the behavior of the REST API.

This is a part of story #13463: Split PATCH docman_items_type/id route

Change-Id: I231279e7423dc487e49581425a6cc2789f67a285