request #13694 Get rid of RabbitMQ
Redis is the only queue system supported.

From a usage point of view, the svnroot_updater script is gone in favor of
`tuleap worker:svnroot-update`.

Most of the fixes are in distlp tests because they were working by accident
(rabbitmq was not actually used for message passing).

This patch doesn't remove the dependency on php-amqplib/php-amqplib yet in composer
because it's too much PITA for rebases. I'll do it in dedicated patchset once this
one is merged.

Change-Id: I2224e4fed517af77720d35052a78ce025229b060