Use Timeframe semantic in Burnup Daily
Context
=======

Burnup Daily event is used to compute burnup "yesterday" data for
current burnup charts. By "current" we mean that end_date (i.e.
start_date + duration) of the milestone is greater than yesterday.

How to test
===========

Have a milestone that is arleady done. Every time you run daily
event[0], then the computation of its burnup chart should not appear in
the backend syslog[1].

Now for the milestone:
- add another date field in tracker administration
- add another duration field in tracker administration
- update manually the database so that tracker_semantic_timeframe points
to those fields for the given tracker
- update milestone artifact to set "another date field" and "another
duration field" values so that computed end date will be greater than
today. Example:
  + "another date field" value = two days ago
  + "another duration field" value = 10

Run daily event. You can assert that burnup chart computation appears in
the backend syslog, meaning that timeframe semantic is taken into
account.

Implementation notes
====================

The subquery used to search field names defined by the semantic is doing
a full table scan of tracker_semantic_timeframe. However, during daily
event, it is not a really a big deal since we are searching for every
artifacts with a burnup on the platform.

---

[0] issue the following commands:
  /usr/share/codendi/src/utils/php-launcher.sh /usr/share/codendi/src/utils/codendi_daily.php
  tuleap process-system-events default
[1] /var/log/tuleap/codendi_syslog

Change-Id: I8218c1fba89f8292e5d5283cebfa3ca1acfb730d