Refactoring: Unit-test DynamicUGroupMembersUpdater
Part of story #13490 have a consistent management of project & groups
members

How to test:
No functional changes expected.

When you add a user to any of the dynamic user groups:
- Project administrators
- Wiki administrators
- Forum administrators
- News writers / editors
- News administrators

Then the user is also added to project members.
When you remove a user from any of those groups, then the user is not
removed from project members.

Change-Id: I6e7b86d274e627651cca55ad3328ae9cfed39bf5