Folders can be copied via the REST API
The REST route to create a new folder now also allow users
to create a new folder from an existing one.

The document UI and the usage of the copy via the REST route
make easy to find out with the copy of links. A copied link
is corrupted. This issue is going to be fixed independently.

Part of story #12456: Cut/Copy/paste items

Change-Id: I9ac3db643f69de0262d283c01065b5b05829a69c