Copy of document is rejected when a file is being uploaded with the title in the destination folder
This work the same way than for the creation of document.
Updating the title of the copied document to "Copy of <title>" would
probably cause more confusion then just rejecting the paste since
there is not yet (and maybe the upload will never be finished) another
document visible with the same name in the folder.

Part of story #12456: Cut/Copy/paste items

Change-Id: I20e5055471b20d9a6df9003c464a45c9c3dd2c21