story #13628 Get only open release (not all)
Instead of get all milestones, and parse it to get only Future Milestone,
it's better to get only open milestones, to treat less milestones.

If a Milestone has no status and no date, it will a current milestone.

And the API return the total size of the request, not just those showed in API.

Furhthermore, I need to recursiveGet (instead of get) the API to get all milestones.

Change-Id: Ic18d151841ec9eac78b822d27fd369b084998b81