Add to project members in bound ugroups
Part of story #13490 have a consistent management of project & groups
members

How to test:
- Set up a ugroup in a Parent project. Add members to the ugroup
In a Child project:
- Create a ugroup.
- When you bind that ugroup to the Parent project/ugroup, then all
members of the ugroup are duplicated in the Child project. All members
that were not project members are now project members of Child project.
- After binding, when you add a new member to the Parent ugroup, a
System Event is dispatched. When processed, the new member becomes
member of the Child project and ugroup.

Given the Forge allows restricted users and the Parent project also
allows them. When you add a restricted user to the Parent ugroup,
- They will also become project member of projects that allow restricted
users (Public, Public incl. Restricted, Private incl. Restricted)
- They will NOT be added to the Child ugroup, and will NOT become
project
member of Private projects.

Note that "Welcome to the project" emails are NOT sent in these
automatic operations.

Change-Id: Ib11cbd792fce67a993532851b7ae9be8a2309d89