Fix the data binding between the GlobalMetadata component and its children
This is a part of story #13589 Update hardcoded metadata

The bad call of IsInUpdateContext props of GlobalMetadata component broke the title check.
When you wanted to update an item via the new update modal the title check always returned an error. Even if the itemtitle is the same title of he updated item.

How to test:
 - Update a document tile via new update moddal
=> if the title is the same than an other document => Error
=> if the title is the same than your current document => No error

Change-Id: I4e5d9b681117c42b07f2c873e6a17583a3b2bfa6