story #13632 Haven't closed milestone in What's Hot/What's Next section in Agile dashboard
In Agile Dashboard, when a release was closed in status semantic and if it has no dates,
so it appeared in the 3 sections (Done, Whats's hot, What's Next).

Now,it appears only on Done section.

How to test:
	- Have a release with no dates and status semantic "closed" => it appears only in Done section
	- Have a release with no dates and status semantic "open" => it appears in What's Hot and What's Next section
	- Have a release with no dates and no status => it appears only in What's Hot section

(I can't add some tests in this files, because it needs a PDO connection)

Change-Id: Ia51e2e12a2005bd52f1a34f276f41f53e60a129a