Inherit synchronized membership setting at project creation
Part of story #13490 have a consistent management of project & groups
members

How to test:
Create a new public project.
- When you choose a public project that has the "Synchronized Project
membership" setting (forge your DB), the setting is inherited as
activated.
- When you choose a public project that does not have the setting, no
row is added in the DB, the setting is not inherited.

When you create a private project, the setting is not duplicated as
private projects always have "synchronized project membership".

Change-Id: Ie03a4514fd35914a7322df894f588d59ce436925