Creation of a folder via the REST route should not check for ongoing file uploads
A folder and a document can cohabit with the same title, as such there is
no reason to reject the creation of a folder if a file with the same name
is being uploaded.

The creation of a new folder was the only spot checking this, other
ways to update the title (metadata update, new version, copy…) do not
have it.

Part of story #12456: Cut/Copy/paste items

Change-Id: I004c5a47c83b3472edfcf950c690aaa0c7aad0b2