Update folder metadata modal
part of story #13589 Update hardcoded metadata

How to test:
 - update a folder:
   => you should be able to choose the recursivity options for its children

I'm sorry about this patch size, but I had to split the status data binding management
Previous one file handle data-binding for item/folder and for create/update
As they each have their own data structure and validation, it's easiest
to understand if the data-binding is done for each particular case.

This patch does not deal with css, it will be done later.

This patch does not correct the fact that default are not applied in create modals
It's broken since a while, and it's not the aim of this particular commit to fix it

The update of current folder will also be done in a dedicated commit.

Change-Id: I76119d8237145c2a88712bc606f5c1569921cc67