request #13644: FRS plugin stylesheets can not be built in a modern dev environnement
The version of node-sass pulled by the FRS plugin is too old
to be usable with node 1.15 and node-gyp can not rebuild the
module because it needs Python 2 to do so.
Anyway there is no need to pull a specific version of node-sass
in this plugin. The same version than for the rest of the build
should be used.

Change-Id: I2de9548cf3a41d8446e161088c90ef74c5d197c4