Replace event management based on CustomEvents by a Vue-based event bus
This approach brings a few advantages:
 * It makes sure that two different instances of the app do
not interact together, the events are kept within the app. This
might becomes handy if the document application needs to be put
in a widget.
 * IE11 is reported by both MDN and Can I Use as not supporting
CustomEvent constructor [0][1].
 * The events can be observed from the Vue dev tools.

It was one of the migration path proposed by the Vue team when
`$dispatch` and `$broadcast` were removed from Vue [2].

No functionnal change.

Part of story #12456: Cut/Copy/paste items

[0] https://github.com/mdn/browser-compat-data/blob/65740c63cb8c861f4511f29b3fa3cb75365f6878/api/CustomEvent.json#L113
[1] https://github.com/Fyrd/caniuse/blob/cdb9d8788bfd147b86bbd1abe3ba3198a2d284ef/features-json/customevent.json#L358
[2] https://vuejs.org/v2/guide/migration.html#dispatch-and-broadcast-replaced

Change-Id: I282157bfb6988d4a4e9afbef22fef0e9fd308cd2