Link's approval table can be mixed up with the approval table of files
The approval table DB table does not distinguish between the version ID
of files and embedded files and the version ID of links. Unfortunately,
Files and links does not use the same version ID key and as a result
a same version ID value can match both a file and a link so a same
approval table can become shared between two different items.

In the situation where the both a file or a link can be the target of
an approval table, this contribution marks the table as corrupted and
a warning message is displayed.

request #13118: approval tables can share same id regardless of their types

Change-Id: If076a81b575bd92863254acf4bf2771e5a912d30