Drop ext-gettext from require key of src/composer.json
The dep has been added by manual edit of the composer.json
in 6ffc0e34e4d1ceb3d55947325e206b8907b67715.
In the Tuleap case it is more annoying than anything else as we are not building
a library and the extension dependencies are managed via RPM packages.

Having the extension as a dependency is annoying for people just wanting to build
the source that will be executed inside an environnement close to production via
Docker containers.
Only people running the unit tests on directly on their host needs to have the extension
installed and loaded, if they don't the tests will crash anyway.

Part of request #13615: Artifact import CSV: warn people about artifacts which are
not able to be imported

Change-Id: Ifb593da283cd38add010e88bfcd5b9469f80e206