Persist document clipboard between page refreshes
The sessionStorage is used to maintain the state between refreshes.
This is preferred over the localStorage to not have to deal with long term
cleanup and limit the exposure of the information stored in it.

Sharing the clipboard between browser tabs will be managed in a ulterior
contribution.

Part of story #12456: Cut/Copy/paste items

Change-Id: I1e92893e34d54c39b85b69b8c8044531263802ed