Fix eslint errors
part of request #13545: Introduce eslint to the test pipeline

This fixes all eslint errors in testmanagement plugin.

Many of the lodash errors have been ignored because they are applied on
objects and not arrays, so there is no "one-liner" native replacement.

How to test:
- npm run eslint -- --quiet ./plugins/testmanagement
- The previous command should return 0 error. "quiet option" disables
warnings, which are not handled by this commit.
- No functional change expected

Change-Id: I0bd4c5d922e32f94ad81bf31c638812ccc8b9f9a