Share clipboard mutations cross-tabs
To make sure the clipboard state is available when the store is initialized,
the data are now stored in the local storage with an expiration delay.

In a ideal world we would have only one plugin to share the mutations and persist
the state but this combination seems to do the trick.

Part of story #12456: Cut/Copy/paste items

Change-Id: I9486dd66bfe3f4adbabe8483975f50be1588f6ed