Convert release widget to Typescript
Part of request #13702 Introduce Typescript in Vue apps

This contribution's scope will limit to introducing Typescript and
making it behave with our build tools.
Another patch will be needed to actually make use of typings in
Typescript and to enforce it.

How to test:
- "npm run build" succeeds
- No functional change in the Release widget app
- "npm run test" succeeds, all tests should pass
- "make generate-po" should succesfully extract translated strings from
release-widget Vue components
- "npm run eslint ." should not take 20min and should not report any
error.
- eslint should report errors in Typescript files
- prettier should report misformatted Typescript files

Note to reviewer:
- "npm run coverage" is known to produce misaligned coverage in this
patch. A further contribution will fix that, but since we do not provide
this coverage to CI for now, I chose to live with it temporarily.

Change-Id: I344f13e05276dc43e1a926078b0313e5fbad1e2f