request #13678: Update to Psalm 3.4.10
Psalm and its dependencies are updated to the latest versions.

This version comes with a nice performance improvement, better
management of the templated types and bugfixes.

New issues found by this version are fixed and the baseline file
is updated.

Also took the opportunity to run Psalm on 2 threads by default
now that we have solved most (hopefully all) inconsistencies.

Change-Id: I7586088f933f7214d9e1dc9ca48be60242bc8f4d