add progress bar
This is part of story #13297 be able to see the progression of my release

Add a progress bar under remaining effort (points and date).
This progress bar represents the percent in the tooltip. Like points or date, if it can't calculate the remaining effort, the color is disabled (grey).

How to test:
	- Install plugin release widget
	- Have some release with:
		- Start date and end date
		- Start date and no end date
		- No start date

		- Remaining effort and initial effort
		- Remaining effort and no initial effort
		- No remaining effort and initial effort
		- No remaining effort and no initial effort

Expected results:
	- Represents the message in the tooltip
		- If there is a percent, the progress bar is displayed in color 'danger' (date) or color 'success' (points)
		- If the remaining effort is done, the color is the color of the theme
		- If it can't calculate the percent, the color is disabled (grey)

Change-Id: I9f35f674a6c843da2c291472448f81e9ad05e2f3