Refactoring: extract frs resources management from core
It used to be done in Core because we thought we had no alternative.
But it proved to be false as we already do it for baseline and docman.

Will make introduction of /project/:id/frs_service easier.

No functional changes expected except some types are now shown in
API explorer.

Part of story #13654 have detailed permissions in REST API

Change-Id: If58e5f77f10dd74b1e68c5cca52096c1f4482172