Reenable Vue rules and fix eslint errors in Vue components
Part of request #13706 eslint should also analyse Vue files

The following command should not report any error

$ npm run eslint -- --quiet .

I took the opportuniy to reenable some rules that were only broken
because of a prospective change in vue-eslint-parser. We never updated
it for that reason, so there's no reason to keep the rules disabled.

Change-Id: Ib74979cc02c4f711508e5737a88110bfa611766a