Add service level permissions for the document manager
Similarly to 0fe405a3f7035bbf4e2b85ad1726ab5bde37ef5a, a new
project level route /porjects/:id/docman_service is introduced.

In addition to the service level permissions, this new route also have
information about the root item of the document manager in the project.
This information is also currently given in the additional_informations
key of of the Project representation. This raise concerns about performances
and usability as the user as no way to know what's will be added into this key.
A breaking change will be introduced to remove the docman informations from
the additional_informations key and resolve those concerns.

Part of story #13654: have detailed permissions in REST API

Change-Id: I0de94addaf7590e413ffa7f7427e64a26e65f894