Expiration date information is returned by REST route
This is part of story #13667 have an expiration date associated to API AccessKey

If you have manually set an expiration date, this information will be
returned by the REST route GET users/:id/access_key.
Expired keys are not returned in the result.

In any other cases, this key (expiration_date) will be NULL.

Change-Id: Id93c2464b5d8911ff828c1cd52f66974dd1d6d6a