request #13726: Cleanup information given by the docman plugin in the REST route GET /project/:id
Some information related to the docman plugin are given in the key additional_information.
This raise concern about the usability of the REST route since there is no way to know what will be
present in additional_information and how it will be formatted. It also not nice for the performance
of the REST route as we end up with information that is, most of the time not needed.

This is a breaking change. However, it was mostly undocumented and users always needed
to check if the information was present or not.

Change-Id: I38aed3eea19a9a1dd5feed916af0554b7bd0e147