Transform Vue.extends to @component
This is part of story #13660 use typescript in the widget

After introduce typescript in plugin release widget, I change all Vue.extend to @component and class.

That's a good practice of Vue with typescript.

When the @component is empty (there aren't component call), it's mandatory to keep '@component' before the class.

No expected functional changes in this patch.

If a method in one class begins with 'get ...', that's a computed method in Vue.
If method haven't 'get' at the begining, that's a method in Vue.
A variable in a class represents some datas in Vue.

Change-Id: Ic327fd9f2ca4f8f973f6ca0ca23be4b9a72e6303