Expiration date can be set by REST
This is part of story #13667 have an expiration date associated to API AccessKey

The REST route POST /access_keys allows now an expiration_date parameter
to set an expiration date for a newly created API key.

Example of JSON content:
{
 "description": "This is my API key",
 "expiration_date": "2019-08-07T10:15:57+02:00"
}

The expiration_date is optional.
If provided, it must be formatted as an ISO-8601 date.
In addition, you cannot create an already expired access key.

Change-Id: Ida4dbadf2191f492c7ad1744f660d801f63b3e0b