Change rules no-unused-var to @typescript-eslint/no-unused-vars
This is part of story #13660 use typescript in the widget

To introduce TypeScript in Tuleap, we need to change one rule to disallow unused vars in Js and Ts.

So, this patch changes this rule and use @typescript-eslint/no-unused-vars which can be use for TS and JS.

No expected functionnal changes.

Change-Id: Ia4e364572929c849001f8a4ab626f7adde8b36c4