Prepare the extraction of the code used to set permissions on docman items
This is a technical change, no functionnal change is expected.

This contribution reworks a bit the code related to the permissions
update. It does not directly depends on a HTTP request anymore which
will the code extraction.

This section of the code needs to be extracted so it can be reused by
the upcoming REST routes.

Part of story #12457: New document modal: permissions

Change-Id: I3b6dc8f73c08d789b50f6fe02f461e40c5b15adb