Display expiration date in user preference page
This is part of story #13667 have an expiration date associated to API AccessKey

A new column (Expire on) is added in user preference page to display API
access key expiration if any (otherwise - is displayed).

If a key is already expired and not yet deleted, then this key is no
more listed in the page.

Change-Id: I830bcb691d8066cba1f6f59916719bcfb3a4c106