Preview the label in the Edit modal
Part of story #13566 associate a custom icon to external services

How to test:
As project admin or site admin, when you edit a service, there is a
dynamic preview of the sidebar link with the label you enter.

When the label is long (up to 40 char), there is an ellipsis at the end,
and it is also in the real sidebar.

Change-Id: I19746881ad3f0fcca2fbb37d343c41969060cc23