request #13739: All the dates associated with access keys does not take into account the user's timezone on the account page
PHP ignores the timezone when a UNIX timestamp is given to the constructor
of the DateTime class [0].

It's not an issue when you only do comparisons with the DateTime object
because a UNIX timestamp represents a unique point in time whatever the
timezone. However, it's an issue when you want to format the date to display
it because the date can seem incorrect for the user.

[0] https://www.php.net/manual/en/datetime.construct.php

Change-Id: I4a275923d4abbbd97ccf04571255830624c3db06