Make sure that all imported JS/TS dependencies are declared in package.json
This solves "request #13724: document does not import sprintf-js properly"
where it was found that sprintf-js was not present in the package.json
of the document plugin. The build was working due to the way Webpack
do its modules resolution [0]. In this case the missing dependency was
found in the root node_modules because one the Tuleap dev dependency
also needs it (in fact it's needed by a package at depth 3 in the
dependency tree).

This unexpected resolution makes dependencies hard to track and can lead
to subtle breakage that can also be hard to debug.
In order to not find ourselves again in the same situation, a new ESLint
plugin is added. Only the rule dedicated to find this issue is enabled
but it comes with other nice rules that we may wish to use at some
point.

[0] https://webpack.js.org/configuration/resolve/#resolvemodules

Change-Id: I8393ffd411504ee94e211345e430eaad32cc216d