request #13753: Use Lerna to install the JS dependencies of all Tuleap components
This is more of a real life experiment than a real necessity right now.
The goal is, for now, to see if how it behaves in the build pipeline and
in the development environements.

Dependency hoisting is explicitely disabled for now, it would requires
some (minimal) changes in the way we work with the dependencies and some
special handling for the plugins maintained outside the main Tuleap
repositories.

Ultimately, we could replace our remaining gulpfile with it.

Adventurous Tuleap developers with some knowledge of the build system
can start experimenting with it. Here are some examples of what you can
do with it (everything is supposed to be executed from the root folder
and prefixed with ./node_modules/.bin/ or npx):

 * lerna run build # Execute the build script of all components, you
might want to add --concurrency X
 * lerna run watch --no-bail --stream # Execute the watch script of all
components
 * lerna clean # Remove the node_modules directory of all components

Change-Id: Id91a839ee6b12e918cc31df9328a4e599e6e71a6