Enforce Typescript type checking
Build of Typescript apps must fail must not skip typechecking.
The transpileOnly mode was nice when introducing Typescript into
the stack but it should not be kept.

This seems to have been missed in 5834c0b94be6e86448a17ded09a2cc810cd3d375.

This is part of story #13660 use typescript in the widget

Change-Id: I2d79ace83639c21c42e66229162cc71d51e25e04