"Default values" should not be used when returning permissions information about a docman item
The document manager does not use those to check if the user has enough
right to do a specific action. When no user group is set for a type of permissions
it means that no user can do the actions (expect if the user is also in a group
higher up in the hierarchy).

Part of story #12457: New document modal: permissions
Part of story #13654: have detailed permissions in REST API

Change-Id: If200109ed630f67c8785fc0f2cdb83971fa560b9