Remove unnecessary null checks in Timeframe Semantic
This change is part of story #12537

Due to an issue with Psalm we introduced redundant null checks
(https://github.com/vimeo/psalm/issues/1994). Now that we know it
actually is a bug and are waiting for the release of the patch, a
@psalm-suppress comment has been added referencing said issue. Once
Psalm is updated, the suppress can be removed and it should still work.

Change-Id: I0384b17f7b3d064042cd2c44d2445910547eb8b8