bugs #13744 Add pagination header if there isn't planning
When we request the API and want all the milestone of a project, and if there isn't milestone in this project, so an exception is throwed.
But when there is this exception, the pagination in the header doesn't exist, and in certain cases (when we use recursiveGet), we need to have the pagination in the header.

How to test:
	- Have a project without agiledashboard
	- Use API with project/milestone

Expected Results:
	- have the pagination in the header (x-pagination-limit, x-pagination-limit-max, x-pagination-offset, x-pagination-size)

Change-Id: I9e87dbcacf18dd8ca75bd5581ea153d7bb826fc5