bugs #13745 Add pagination header if there isn't backlog
When we request the API and want items in the backlog of a project, and if there isn't backlog in this project, so an exception is throwed.
But when there is this exception, the pagination in the header doesn't exist, and in certain cases (when we use recursiveGet), we need to have the pagination in the header.

How to test:
	- Have a project without agiledashboard
	- Use API with project/backlog

Expected Results:
	- have the pagination in the header (x-pagination-limit, x-pagination-limit-max, x-pagination-offset, x-pagination-size)

Change-Id: I650aa800554e0b7876b35f78eb8dcff39d9166aa