Add $end_date property in TimePeriodWithoutWeekEnd
Instead of having to compute end date from start date and duration, we
now store directly the real value of start date, end date, and duration.

This allows us to have a more consistent TimePeriod in the case of end
date < start date for example (negative duration instead of random
value).

This will also allow us to instantiate weird TimePeriod based on a end
date but no start date for example. I don't really know the meaning of
this, but it will be consistent with start date/duration behavior.

Part of story #12537: configure my burndown/up & sprint with a start date & end date

Change-Id: I5e47b7312e22b81b953779381dc2b580797121dd