Fix "is_required" key value of the Metdadata RESTs representations
This is a part of story #12634 Add custom metadata in modal

In the REST representation of the metadata the value of the "is_required" key is always false.

Why?
The class attribute $isRequired of the Docman_Metdata.php class is always false.
When the metadata is stored in the database the 'required' column of plugin_docman_metadata table is always false! (= 0)
It's always false (=0) because at the metadata creation we always init the row 'required' parameter to false  ($mdrow['required'] = false)
See Docman_Actions.class.php L 1253

To know if a metadata is required or not we had to rely on $isEmptyValueAllowed from the Docman_Metadata.php

You can the test fix via the API explorer:
 => GET /projects/docman_metadata
 => GET /docman_items/:id
=> When a metadata is required the "is_required" should be true

You can also test via the new document UI:
 => When a custom metadata is required an asterisk should be displayed next to the label

Change-Id: I93ba93b57d53a1646933e92a1eadf8716d9dc2f2