Display burnup with count elements values
This is part of story #13735 have burn-up based on item count

You have to configure your Tuleap to be able to display burnup with
count elements:

* use_burnup_count_elements set to 1 to enable the feature
* burnup_ids_count_elements is a comma separated value with the
  milestone ids you want to use count elements for burnup.

The points representing count elements are always added into the Burnup
representation.

Given the burnup mode, either effort or count elements is used.

Chart labels are not yet updated. This will be done in a dedicated
commit.

Change-Id: I72ba2b9fa89cf5379734588970a9b32497d754b2