request #13747: REST endpoint GET /projects/:id/user_groups returns system user groups with a not recognized ID
The IDs of the system user groups are accepted by the REST endpoint
GET /user_groups/:id.
The identifiers <project_id>_<system_user_group_id> that were accepted
(but not documented) are now rejected.

The endpoints allowing the update of permissions of a document manager
item are now accepting the system user groups.

Change-Id: Iaf81dfd67b06c490bd839af8867eda296e87c1a1