story #13754 Build translation TypeScript and Vue in same time
It's a problem encountered when a Vue app needs to use some functions in
ts files.
And in these ts files there are translations.
At the build, we can't use at the same time 'rule_easygettext_loader'
and 'rule_po_file'.

So with this patch, we can build all translations (in Vue + in ts file)
with only 'rule_easygettext_loader'.
To do it, it needs to use gettext of vue.

How to test:
	- Have a vue app to call function in ts file (you can pass in
argument a HTMLElement of your template to assign the message).
	- Have a ts file with vue gettext (the variable must be called
'gettext_provider')
		Example:
			let gettext_provider = new Vue();
			gettext_provider.$gettext("My message");

	- Have in build-manifest, 'gettext-vue' object
	- Have 'rule_easygettext_loader' in build-
	- Run 'make generate-po'

Expected result:
	- In po file, find "My message"
	- Build the app, and see the translation

Change-Id: I254a26564b176e550dcd44113edee5b1989e2d8d